created helper file for book actions, moved files
