+++ src/configuration/axios.js
@@ -1,6 +1,6 @@
 import axios from 'axios'
 
-import { API_BASE, API_ENDPOINTS } from "./constants";
+import { API_BASE, API_ENDPOINTS } from "@/constants";
 
 const userRoute = API_BASE + API_ENDPOINTS.user
 const booksRoute = API_BASE + API_ENDPOINTS.books
