+++ src/main.js
@@ -1,6 +1,6 @@
 import { createApp } from 'vue'
 import App from './App.vue'
-import router from "@/router"
+import router from "@/router/router"
 import 'bootstrap'
 import 'bootstrap/dist/css/bootstrap.min.css'
 import { createPinia } from 'pinia'
