+++ src/store/modules/books/actions.js
@@ -1,12 +1,12 @@
-import {book_api} from "@/axios";
+import {book_api} from "@/configuration/axios";
 import {Book} from "@/models/Book";
+import {bookActivation, cancelReservationHelper, receiveBookHelper, reserveBookHandler} from "@/utils/api-utils"
 
 export default {
     async fetchBooks() {
         try {
             const response = await book_api.get('/fetch_books');
             const data = await response.data
-            console.log(data)
             this.bookList = data.map(bookData => new Book(bookData));
         } catch (error) {
             console.log(error)
@@ -39,13 +39,7 @@ export default {
             const response = await book_api.patch(`/reserve_book/${id}`)
             if (response.status === 200) {
                 const responseData = response.data.data;
-                for (const book in this.bookList) {
-                    if (book.id === responseData.id) {
-                        book["reserved"] = true;
-                        book["lenderId"] = responseData.lenderId;
-                    }
-                }
-                return true;
+                reserveBookHandler(responseData, this.bookList);
             }
         } catch (error) {
             console.log(error)
@@ -55,14 +49,7 @@ export default {
         try {
             const response = await book_api.patch(`/cancel_reservation/${bookId}`);
             if (response.status === 200) {
-                for (const key in this.bookList) {
-                    const book = this.bookList[key];
-                    if (book.id === bookId) {
-                        book["reserved"] = false;
-                        book["lenderId"] = null;
-                    }
-                }
-                return true;
+                cancelReservationHelper(bookId, this.bookList);
             }
         } catch (error) {
             console.log(`Failed to cancel reservation book id: ${bookId}`, error)
@@ -73,16 +60,7 @@ export default {
             const response = await book_api.patch(`/receive_book/${bookId}`);
             if (response.status === 200) {
                 const returnDateStr = response.data.returnDate;
-                const [day, month, year] = returnDateStr.split("-");
-                const returnDate = new Date(year, month - 1, day);
-                for (const key in this.bookList) {
-                    const book = this.bookList[key];
-                    if (book.id === bookId) {
-                        book["lentOut"] = true;
-                        book["returnDate"] = returnDate;
-                    }
-                }
-                return true;
+                receiveBookHelper(returnDateStr, bookId, this.bookList)
             }
         } catch (error) {
             console.error(`Cannot 'mark as received' book id ${bookId}`);
@@ -90,9 +68,7 @@ export default {
     },
     async returnBook(bookId) {
         try {
-            console.log(bookId)
             const response = await book_api.patch(`/return_book/${bookId}`);
-            console.log(response.status)
             if (response.status === 200) {
                 for (const key in this.bookList) {
                     const book = this.bookList[key];
@@ -113,13 +89,7 @@ export default {
             const response = await book_api.patch(`/activity/${bookId}`);
             if (response.status === 200) {
                 const responseData = response.data.data;
-                for (const key in this.bookList) {
-                    const book = this.bookList[key];
-                    if (book.id === bookId) {
-                        book["isActive"] = responseData;
-                        return true;
-                    }
-                }
+                bookActivation(responseData,bookId,this.bookList);
             }
         } catch (error) {
             console.error(error);
